refactor: split playUtils functions
